alignment/asmd/pyproject.toml
[tool.poetry]
name = "asmd"
version = "0.3"
description = "Audio-Score Meta-Dataset"
authors = ["Federico Simonetta <federico.simonetta@unimi.it>"]
readme = "README.rst"
homepage = "https://asmd.readthedocs.org"
documentation = "https://asmd.readthedocs.org"
repository = "https://gitlab.di.unimi.it/federicosimonetta/asmd/"
license = "MIT"
classifiers = [
    "Topic :: Multimedia :: Sound/Audio",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Intended Audience :: Other Audience"
]
build = "build.py"

[tool.poetry.dependencies]
python = "^3.8.5"
setuptools = "^45.2.0"
numpy = "^1.18.1"
requests = "^2.22.0"
beautifulsoup4 = "^4.8.2"
plotly = "^4.4.1"
pretty_midi = "^0.2.8"
scikit_learn = "^0.22.1"
pyfiglet = "^0.8.post1"
alive-progress = "^1.3.3"
prompt_toolkit = "^3.0.3"
decorator = "^4.4.1"
essentia = "^2.1b6.dev374"
"mega.py" = "^1.0.6"
joblib = "^0.14.1"
tqdm = "^4.43.0"
hmmlearn = "^0.2.5"

[tool.poetry.dev-dependencies]
ipdb = "*"
ipython = "*"
cython = "^0.29.21"

[build-system]
requires = ["poetry>=0.12", "cython>=0.29.14"]
build-backend = "poetry.masonry.api"
